create utility function for getting access & refresh token & setting browser cookies
