+++ src/app/modules/auth/auth.controller.ts
@@ -5,6 +5,7 @@ import { sendResponse } from "../../utils/sendResponse";
 import httpStatus from "http-status-codes";
 import { AuthService } from "./auth.service";
 import AppError from "../../errorHelpers/AppError";
+import { setAuthCookie } from "../../utils/setCookie";
 
 const credentialsLogin = catchAsync(
   async (req: Request, res: Response, next: NextFunction) => {
@@ -13,15 +14,17 @@ const credentialsLogin = catchAsync(
     const loginInfo = await AuthService.credentialsLogin(req.body);
 
     // after done getNewAccessToken's works
-    res.cookie("accessToken", loginInfo.accessToken, {
-      httpOnly: true,
-      secure: false,
-    });
-    res.cookie("refreshToken", loginInfo.refreshToken, {
-      httpOnly: true,
-      secure: false,
-    });
+    // res.cookie("accessToken", loginInfo.accessToken, {
+    //   httpOnly: true,
+    //   secure: false,
+    // });
+    // res.cookie("refreshToken", loginInfo.refreshToken, {
+    //   httpOnly: true,
+    //   secure: false,
+    // });
 
+    setAuthCookie(res, loginInfo);
+    
     sendResponse(res, {
       statusCode: httpStatus.OK,
       success: true,
@@ -42,6 +45,12 @@ const getNewAccessToken = catchAsync(
       );
     }
     const tokenInfo = await AuthService.getNewAccessToken(refreshToken);
+
+    // res.cookie("accessToken", tokenInfo.accessToken, {
+    //   httpOnly: true,
+    //   secure: false,
+    // });
+    setAuthCookie(res, tokenInfo);
 
     sendResponse(res, {
       statusCode: httpStatus.OK,
