+++ src/app/utils/userTokens.ts
@@ -1,7 +1,12 @@
+import { JwtPayload } from "jsonwebtoken";
 import { envVars } from "../config/env";
-import { IUser } from "../modules/user/user.interface";
-import { generateToken } from "./jwt";
+import { IsActive, IUser } from "../modules/user/user.interface";
+import { generateToken, verifyToken } from "./jwt";
+import { User } from "../modules/user/user.model";
+import AppError from "../errorHelpers/AppError";
+import httpStatus from "http-status-codes";
 
+//createUserTokens
 export const createUserTokens = (user: Partial<IUser>) => {
   // JWT
   const jwtPayload = {
@@ -26,4 +31,48 @@ export const createUserTokens = (user: Partial<IUser>) => {
     accessToken,
     refreshToken,
   };
+};
+
+
+// createNewAccessTokenWithRefreshToken
+export const createNewAccessTokenWithRefreshToken = async (
+  refreshToken: string
+) => {
+  const verifiedRefreshToken = verifyToken(
+    refreshToken,
+    envVars.JWT_REFRESH_SECRET
+  ) as JwtPayload;
+
+  const isUserExist = await User.findOne({ email: verifiedRefreshToken.email });
+
+  if (!isUserExist) {
+    throw new AppError(httpStatus.BAD_REQUEST, "User Doesn't Exist!", "");
+  }
+  if (
+    isUserExist.isActive === IsActive.BLOCKED ||
+    isUserExist.isActive === IsActive.INACTIVE
+  ) {
+    throw new AppError(
+      httpStatus.BAD_REQUEST,
+      `User is ${isUserExist.isActive}  !`,
+      ""
+    );
+  }
+  if (isUserExist.isDeleted) {
+    throw new AppError(httpStatus.BAD_REQUEST, "User is Deleted!", "");
+  }
+
+  // JWT
+  const jwtPayload = {
+    userId: isUserExist._id,
+    email: isUserExist.email,
+    role: isUserExist.role,
+  };
+
+  const accessToken = generateToken(
+    jwtPayload,
+    envVars.JWT_ACCESS_SECRET,
+    envVars.JWT_ACCESS_EXPIRES
+  );
+  return accessToken;
 };
